Revert "Simplified significantly to mainly to make star viewer faster" Goal is making an SQL database, NOT  star viewer This reverts commit 651d0918f63ccc05b2fda6b6e3b93f2ae06eb70d.

+++ client/src/client.h
@@ -253,9 +253,9 @@ int64_t vdbreader_next_int(struct VdbReader* r) {
     return i;
 }
 
-float vdbreader_next_float(struct VdbReader* r) {
-    float d = *((float*)(r->buf + r->idx));
-    r->idx += sizeof(float);
+double vdbreader_next_float(struct VdbReader* r) {
+    double d = *((double*)(r->buf + r->idx));
+    r->idx += sizeof(double);
     return d;
 }
